subbort images

+++ utils/file.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-function deleteAllFilesFormOS(req) {
+function deleteAllFilesFormOS(req, forImage = false) {
 
     fs.rm('data/imageDir/out', {
         recursive: true,
@@ -21,14 +21,16 @@ function deleteAllFilesFormOS(req) {
         }
     })
 
-    fs.rm(`data/imageDir/${req.file.originalname.replace(".pdf", "")}_page_1.png`, (error) => {
-        if (error) {
-            console.log(error);
-        }
-        else {
-            console.log("File Deleted!");
-        }
-    })
+    if (!forImage) {
+        fs.rm(`data/imageDir/${req.file.originalname.replace(".pdf", "")}_page_1.png`, (error) => {
+            if (error) {
+                console.log(error);
+            }
+            else {
+                console.log("File Deleted!");
+            }
+        })
+    }
 }
 
 module.exports = {
